Improved PX file downloading functionality
The new function download_px() takes in one mandatory argument
and four kwargs with set defaults.

px_objs: A single Px object or a list of Px objects to be downloaded.

target_dir: The directory to which the Px file will be downloaded.
    Relative paths are accepted and are made absolute. Defaults to
    '.'.

compressed: If true, the requests to the remote server set the HTTP
    'Accept-Encoding: gzip' indicating that the returned files should
    be compressed. Downloaded files are decompressed before saving on
    disk. Defaults to False.

sleep: Amount of seconds to sleep between requests to remote server.
    Defaults to 1.

refresh: Determines the policy used to determine whether previously
    downloaded PX files should be redownloaded. Accepted values are
    'never', 'check' and 'always'. Setting this to 'always' results
    in all files being downloaded every time. Setting to 'never' skips
    all files that are already present on disk, without regard for
    file size, contents or timestamp. The setting 'check' fires off
    a HTTP HEAD request to the server and only downloads the file
    if the server's response contains a file timestamp never than
    that of the file on disk. Default to 'check'.

NB: As of writing, the StatFi API does *not* acknowledge the HTTP
header requesting gzipped response in any way, but rather returns
a non-compressed file. It is unclear whether this functionality has
ever truly been supported by the StatFi API. Nevertheless, the
functionality is being kept as it was part of the python2 API.